cli/requirements.txt
google-genai
moviepy
pydub
demucs
google-cloud-storage
pyrubberband
soundfile
click